feat(menu): menu 모듈 분리
- acceptance 요청/확인 메소드 testFixture로 이동
