+++ module-menu/src/test/java/kitchenpos/menu/domain/MenuProductTest.java
