+++ module-menu/src/testFixtures/java/kitchenpos/menu/domain/FixtureMenu.java
