--- module-menugroup/src/main/java/kitchenpos/menugroup/MenuGroupApplication.java
@@ -1,12 +1,0 @@
-package kitchenpos.menugroup;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MenuGroupApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MenuGroupApplication.class, args);
-    }
-}
