+++ module-menugroup/src/test/java/kitchenpos/menugroup/acceptance/MenuGroupAcceptanceTest.java
@@ -5,18 +5,17 @@ import kitchenpos.menugroup.dto.MenuGroupResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DynamicNode;
 import org.junit.jupiter.api.TestFactory;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static kitchenpos.menugroup.acceptance.MenuGroupAcceptanceTestUtil.메뉴_그룹_목록_응답됨;
+import static kitchenpos.menugroup.acceptance.MenuGroupAcceptanceTestUtil.메뉴_그룹_목록_조회_요청;
+import static kitchenpos.menugroup.acceptance.MenuGroupAcceptanceTestUtil.메뉴_그룹_목록_확인됨;
+import static kitchenpos.menugroup.acceptance.MenuGroupAcceptanceTestUtil.메뉴_그룹_생성_실패됨;
+import static kitchenpos.menugroup.acceptance.MenuGroupAcceptanceTestUtil.메뉴_그룹_생성_요청;
+import static kitchenpos.menugroup.acceptance.MenuGroupAcceptanceTestUtil.메뉴_그룹_생성됨;
 import static org.junit.jupiter.api.DynamicTest.dynamicTest;
 
 @DisplayName("메뉴 그룹 관련 기능")
@@ -43,41 +42,5 @@ public class MenuGroupAcceptanceTest extends AcceptanceTest {
                     메뉴_그룹_목록_확인됨(response, "신메뉴");
                 })
         );
-    }
-
-    public static MenuGroupResponse 메뉴_그룹_등록됨(String name) {
-        return 메뉴_그룹_생성_요청(name).getBody();
-    }
-
-    public static ResponseEntity<MenuGroupResponse> 메뉴_그룹_생성_요청(String name) {
-        Map<String, Object> request = new HashMap<>();
-        request.put("name", name);
-        return restTemplate().postForEntity("/api/menu-groups", request, MenuGroupResponse.class);
-    }
-
-    public static ResponseEntity<List<MenuGroupResponse>> 메뉴_그룹_목록_조회_요청() {
-        return restTemplate().exchange("/api/menu-groups", HttpMethod.GET, null,
-                                       new ParameterizedTypeReference<List<MenuGroupResponse>>() {});
-    }
-
-    public static void 메뉴_그룹_생성됨(ResponseEntity<MenuGroupResponse> response) {
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
-        assertThat(response.getHeaders().getLocation()).isNotNull();
-    }
-
-    public static void 메뉴_그룹_생성_실패됨(ResponseEntity<MenuGroupResponse> response) {
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-
-    public static void 메뉴_그룹_목록_응답됨(ResponseEntity<List<MenuGroupResponse>> response) {
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-    }
-
-    public static void 메뉴_그룹_목록_확인됨(ResponseEntity<List<MenuGroupResponse>> response, String... names) {
-        List<String> productNames = response.getBody()
-                                            .stream()
-                                            .map(MenuGroupResponse::getName)
-                                            .collect(Collectors.toList());
-        assertThat(productNames).containsExactly(names);
     }
 }
