--- module-product/src/main/java/kitchenpos/product/ProductApplication.java
@@ -1,12 +1,0 @@
-package kitchenpos.product;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProductApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ProductApplication.class, args);
-    }
-}
