+++ module-product/src/test/java/kitchenpos/ProductApplicationTests.java
@@ -1,0 +1,12 @@
+package kitchenpos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ProductApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+}
