+++ module-test/build.gradle
@@ -5,8 +5,8 @@ plugins {
 }
 
 dependencies {
-    testFixturesImplementation 'org.springframework.boot:spring-boot-starter-web'
-    testFixturesImplementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    testFixturesImplementation 'org.springframework.boot:spring-boot-starter-test'
+    testFixturesApi 'org.springframework.boot:spring-boot-starter-web'
+    testFixturesApi 'org.springframework.boot:spring-boot-starter-data-jpa'
+    testFixturesApi 'org.springframework.boot:spring-boot-starter-test'
     testFixturesImplementation 'com.google.guava:guava:31.1-jre'
 }
